Update to fix Observations
- Fixed issue with UNAVAILABLE being passed through SHDR for DataSet, Table, and TimeSeries
- Added SampleCount property for TimeSeries
- Fixed issue with Count property output for DataSet and Table

+++ src/MTConnect.NET-Common/Observations/ISampleTimeSeriesObservation.cs
@@ -17,7 +17,7 @@ namespace MTConnect.Observations
         /// The number of readings of the value of a data item provided in the data returned when the representation attribute for teh data item is TIME_SERIES.
         /// SampleCount is not provided for data items unless the representation attribute is TIME_SERIES and it MUST be specified when the attribute is TIME_SERIES.
         /// </summary>
-        long SampleCount { get; }
+        int SampleCount { get; }
 
         /// <summary>
         /// Time Series observation MUST report multiple values at fixed intervals in a single observation.
